Fixes #7
--ESDocumentProductCombination.php
Added Ecommerce standards document class that contains a list of product combination records, and a list of product combination profile records

--ESDRecordCombinationProfile.php
Added Ecommerce Standards Record class that holds data for a single combination profile. Each profile defines the fields and field-values that are available for any number of combination products, downloads, or labour.

--ESDRecordCombinationProfileField.php
Added Ecommerce Standards Record class for storing the field data associated to a combination profile. Each field can contain any number of field-values, which allows products, downloads, and labour to be assigned to.

--ESDRecordProductCombination.php
Added Ecommerce Standards record class for associating a child product with a given combination of fields and values

--ESDRecordProductCombinationParent.php
Added Ecommerce Standards record class for associating a parent product with a product combination profile, and all of its associated product-field-value combinations